+++ crystalfontz/__init__.py
@@ -37,6 +37,7 @@ from crystalfontz.keys import (
 )
 from crystalfontz.lcd import LcdRegister
 from crystalfontz.packet import Packet
+from crystalfontz.protocol import ClientProtocol
 from crystalfontz.report import LoggingReportHandler, NoopReportHandler, ReportHandler
 from crystalfontz.response import (
     AtxPowerSwitchFunctionalitySet,
@@ -82,6 +83,7 @@ __all__: List[str] = [
     "BootStateStored",
     "ClearedScreen",
     "Client",
+    "ClientProtocol",
     "Command",
     "CommandSentToLcdController",
     "connection",