+++ CHANGELOG.md
@@ -32,6 +32,7 @@ yyyy/mm/dd Version 2.0.0
 - Byte CLI arguments are validated as being in range
 - Watchdog CLI argument allows "disable" and "disabled" values
 - Configure device model, hardware revision and firmware revison in CLI
+- Do not show stack trace on connection errors in CLI
 
 2025/01/04 Version 1.0.0
 ------------------------